Refactored entities into interface
